cast to fix warning
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@37366 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormViewWithContentViewEditor.m
@@ -324,7 +324,7 @@ int _sortViews(id view1, id view2, void *context)
       prevRect = currRect;
     }
 
-  NSDebugLog(@"The vote is %ld",count);
+  NSDebugLog(@"The vote is %ld",(long int)count);
 
   if(count >= 0)
     vertical = YES;
